Task 3 Update
I updated the Task 3 code a little so it works now. The issue was with trying to create a local 1000-integer array, the Arduino did not seem to like having to do that memory allocation. When I made it a global, the code worked as expected, though I did notice the array took up a lot (like 53%) of the global data space, so that could be an issue in the future.

Arduino sketches also like to be in their own folders, so when I edited it today, it created a new folder and placed the sketch in that folder.