Updated Serial Test
So I tried to commit this from the Raspberry Pi, but I don't think it worked, so I'm committing it from my laptop.

Anyway, this is the uptate that made the Serial testing more reliable between the Pi and the Arduino. I think we can say that we can successfully talk to each other now!