Motor Control
Added a simple algorithm that can spin the motor 1 revolution, provided all the hardware is connected as it should be according to the Motor Specs document.

Also, I think I pulled the origin correctly, but make sure I did before you merge this into master so we don't lose your work on the Raspberry Pi.