Fixed Off-in-The-Weeds Bug and Optimized Manual Turning
This code should be good for the 1st Semester Demo on December 3rd. The bug in the G-Code reader has been fixed to allow for proper conversion to unsigned long ints, so that executing a distance longer than 204 mm does not send 4294967295 pulses to the motor and attempt to send the winding head to Jordan Lake.

The structure to send manual commands to spin the motors once has also been optimized so that different commands exist to move on each axis, and in both directions (rather than having to change the code to switch axes or directions).