G27
Added the G27 command, which indicates that the current position is (0, 0, 0) rather than using G28 for both marking and moving to the (0, 0, 0) point. This also fixes a bug where trying to send the first G28 command after moving causes the same movement to be executed again, as G27 now handles this.