ADC Reference Voltage Experiment
The post-winding test code now contains a proof-of-concept of the algorithm to change the reference voltage of the ADC to 1.1 V for more precise resistance measuring. It will soon contain code that implements it, along with the use of five input resistors and the selection of the best one to use for measuring the resistance.