Rotary Encoder Reading
Reads the rotary encoder and processes the distance of the wire payout. The rotary encoder is only processed when the Arduino is sent the command "(StartEncode)" and stops reading when sent the command "(StopEncode)". At the end of the winding algorithm, the distance in mm is sent to the Pi.